fix: adding LoginElement to island only
